+++ pipelines/input_assembly_info.hpp
@@ -13,7 +13,7 @@ namespace wrapper::directx12 {
 
 		~input_assembly_info() = default;
 
-		input_assembly_info& add_input_element(const std::string& name, const DXGI_FORMAT& format, size_t slot = 0);
+		input_assembly_info& add_input_element(const std::string& name, const DXGI_FORMAT& format, uint32 slot = 0);
 
 		D3D12_INPUT_LAYOUT_DESC desc() const;
 	private:
